introducing rest-dto lib into rest-api and rest-api-client
