+++ apps/java/libraries/case-engine-rest-client/pom.xml
@@ -35,9 +35,16 @@
 
 		<dependency>
 			<groupId>com.wks</groupId>
+			<artifactId>case-engine-rest-dto</artifactId>
+			<version>${project.parent.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>com.wks</groupId>
 			<artifactId>api-security</artifactId>
 			<version>${project.parent.version}</version>
 		</dependency>
+		
 		<dependency>
 			<groupId>org.apache.httpcomponents.client5</groupId>
 			<artifactId>httpclient5</artifactId>
