+++ apps/java/libraries/case-engine-rest-client/src/main/java/com/wks/api/client/gateway/RestTemplateConfig.java
@@ -9,7 +9,7 @@
  * 
  * For licensing information, see the LICENSE file in the root directory of the project.
  */
-package com.wks.bpm.externaltask.api.gateway;
+package com.wks.api.client.gateway;
 
 import java.util.Collections;
