+++ apps/java/libraries/case-engine-rest-client/src/main/java/com/wks/api/client/gateway/impl/CaseInstanceApiGateway.java
@@ -9,14 +9,14 @@
  * 
  * For licensing information, see the LICENSE file in the root directory of the project.
  */
-package com.wks.bpm.externaltask.api.gateway.impl;
+package com.wks.api.client.gateway.impl;
 
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 
-import com.wks.bpm.externaltask.api.gateway.ApiGateway;
+import com.wks.api.client.gateway.ApiGateway;
 
 /**
  * @author victor.franca
