+++ apps/java/services/case-engine-rest-api/pom.xml
@@ -67,6 +67,12 @@
 		</dependency>
 
 		<dependency>
+			<groupId>com.wks</groupId>
+			<artifactId>case-engine-rest-dto</artifactId>
+			<version>${project.parent.version}</version>
+		</dependency>
+
+		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
 			<scope>test</scope>
